made tech and world news dynamic
